==> 1.Joined Dataset.ipynb <==
# Import packages
import matplotlib
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
from pyspark.sql import SQLContext
sqlContext = SQLContext(sc)


df_blood1_raw = pd.read_excel('2. BLOODWORK_blindedSep28.xlsx')  
df_static_raw = pd.read_excel('1. LISTING_blindedSep28.xlsx')
df_blood2_raw = pd.read_csv('cleaned_blood2.csv')

df_static_raw.head() 

df_blood2_raw.head()

df_blood1_raw['SUBJECT_ID'] = df_blood1_raw['SUBJECT_ID'].map(str)

df_blood1_raw.head()

## Data cleaning on blood1 table

# Data cleansing: extract numeric part from cells with mixed values ( 117 norm -> 117)
df_blood1_raw.INR = df_blood1_raw.INR.map(str)
df_blood1_raw.INR = df_blood1_raw.INR.str.extract(r'([-]?[\d.]*)')
df_blood1_raw.CREAT = df_blood1_raw.CREAT.map(str)
df_blood1_raw.CREAT = df_blood1_raw.CREAT.str.extract(r'([-]?[\d.]*)')
df_blood1_raw.BILT = df_blood1_raw.BILT.map(str)
df_blood1_raw.BILT = df_blood1_raw.BILT.str.extract(r'([-]?[\d.]*)')
df_blood1_raw.AST = df_blood1_raw.AST.map(str)
df_blood1_raw.AST = df_blood1_raw.AST.str.extract(r'([-]?[\d.]*)')
df_blood1_raw.ALT = df_blood1_raw.ALT.map(str)
df_blood1_raw.ALT = df_blood1_raw.ALT.str.extract(r'([-]?[\d.]*)')
df_blood1_raw.ALK_PHOS = df_blood1_raw.ALK_PHOS.map(str)
df_blood1_raw.ALK_PHOS = df_blood1_raw.ALK_PHOS.str.extract(r'([-]?[\d.]*)')
df_blood1_raw.PLATELETS = df_blood1_raw.PLATELETS.map(str)
df_blood1_raw.PLATELETS = df_blood1_raw.PLATELETS.str.extract(r'([-]?[\d.]*)')
df_blood1_raw.WBC = df_blood1_raw.WBC.map(str)
df_blood1_raw.WBC = df_blood1_raw.WBC.str.extract(r'([-]?[\d.]*)')
df_blood1_raw.ALB = df_blood1_raw.ALB.map(str)
df_blood1_raw.ALB = df_blood1_raw.ALB.str.extract(r'([-]?[\d.]*)')
df_blood1_raw.EST_GFR = df_blood1_raw.EST_GFR.map(str)
df_blood1_raw.EST_GFR = df_blood1_raw.EST_GFR.str.extract(r'([-]?[\d.]*)')
df_blood1_raw.MELD = df_blood1_raw.MELD.map(str)
df_blood1_raw.MELD = df_blood1_raw.MELD.str.extract(r'([-]?[\d.]*)')
df_blood1_raw.HGB = df_blood1_raw.HGB.map(str)
df_blood1_raw.HGB = df_blood1_raw.HGB.str.extract(r'([-]?[\d.]*)')


df_blood1_raw.head()

df_static_raw['UNLISTED_REASON']

## Combine two blood tables 

#create identifier to join blood tables 
df_blood1_raw["lab_id"] = df_blood1_raw["SUBJECT_ID"].map(str) + df_blood1_raw["LISTING_DATE"].map(str) + df_blood1_raw["DATE_OF_LAB"].map(str)

# rename the columns for the blood2 table 
df_blood2_raw.columns = ['lab_id', 'SUBJECT_ID_NA', 'LISITNG_DATE','DATE_OF_LAB_NA','DAYS_FROM_LISTING','Na','HCO3','GLUCOSE','CALCIUM','MG','PHOS']

df_blood_raw = pd.DataFrame()

# join table blood1 and blood 2
df_blood_raw = df_blood1_raw.merge(df_blood2_raw, left_on='lab_id', right_on='lab_id',how='outer')


#df_blood_raw.to_csv('raw_blood.csv')
df_blood_raw.head()
df_blood_raw.shape

# Preparing joining key for blood and static table 
df_static_raw['LIST_ID'] =  df_static_raw['SUBJECT_ID'].map(str)+df_static_raw['LISTING_DATE'].map(str)
df_blood_raw['LIST_ID'] =  df_blood_raw['SUBJECT_ID'].map(str)+df_blood_raw['LISTING_DATE'].map(str)

#df_blood_cleaned.to_csv('cleaned_blood.csv')
#df_static_raw.to_csv('cleaned_static.csv')

# transfer some columns into string in order to generate sql tables 
df_static_raw['PREV_SURG1'] = df_static_raw['PREV_SURG1'].map(str)
df_static_raw['CAUSE_OF_DEATH_DESCR'] = df_static_raw['CAUSE_OF_DEATH_DESCR'].map(str)
df_static_raw['TREATMENT_PHASE_DESCR'] = df_static_raw['TREATMENT_PHASE_DESCR'].map(str)
df_static_raw['DIALYSIS_TYPE_DESCR'] = df_static_raw['DIALYSIS_TYPE_DESCR'].map(str)
df_static_raw['DIALYSIS_START_DATE'] = df_static_raw['DIALYSIS_START_DATE'].map(str)
df_static_raw['RH_FACTOR_DESCR'] = df_static_raw['RH_FACTOR_DESCR'].map(str)
df_static_raw['SEX_DESCR'] = df_static_raw['SEX_DESCR'].map(str)
df_static_raw['PATIENT_STATUS_DESCR'] = df_static_raw['PATIENT_STATUS_DESCR'].map(str)
df_static_raw['WEIGHT_AT_LISTING'][df_static_raw['WEIGHT_AT_LISTING'] == '76..2'] = 76.2
pd.to_numeric(df_static_raw['WEIGHT_AT_LISTING'])
pd.to_numeric(df_static_raw['HEIGHT_AT_LISTING'])


## Combine static and blood tables 

# Merge the blood table with the static table
left = df_static_raw
right = df_blood_raw
df_combined_raw = pd.merge(left, right, on='LIST_ID')

list(df_combined_raw)
df_combined_raw.shape

df_combined_raw.to_csv('combined_raw.csv')

## Drop and rename colcumns

df_testing = pd.read_csv('combined_raw.csv')

df_testing.shape


df_testing['CREAT'] = pd.to_numeric(df_testing['CREAT'], errors='coerce')
df_testing['CREAT'].dtypes

list(df_testing)

df_testing.drop(['Unnamed: 45','SUBJECT_ID_y','LISTING_DATE_y','SUBJECT_ID_NA','LISITNG_DATE','DATE_OF_LAB_NA'],axis=1,inplace = True)
df_testing.rename(columns={"CALCIUM": "CALC"}, inplace=True)
df_testing.rename(columns={"PREV_SURG1": "PREV_SURG"}, inplace=True)

df_testing.rename(columns={"SUBJECT_ID_x": "SUBJECT_ID","LISTING_DATE_x":"LISTING_DATE"}, inplace=True)

df_testing['BILT'] = pd.to_numeric(df_testing['BILT'], errors='coerce')
df_testing['AST'] = pd.to_numeric(df_testing['AST'], errors='coerce')
df_testing['ALT'] = pd.to_numeric(df_testing['ALT'], errors='coerce')
df_testing['ALK_PHOS'] = pd.to_numeric(df_testing['ALK_PHOS'], errors='coerce')
df_testing['PLATELETS'] = pd.to_numeric(df_testing['PLATELETS'], errors='coerce')
df_testing['WBC'] = pd.to_numeric(df_testing['WBC'], errors='coerce')
df_testing['ALB'] = pd.to_numeric(df_testing['ALB'], errors='coerce')
df_testing['HGB'] = pd.to_numeric(df_testing['HGB'], errors='coerce')
df_testing['INR'] = pd.to_numeric(df_testing['INR'], errors='coerce')
df_testing['CREAT'] = pd.to_numeric(df_testing['CREAT'], errors='coerce')

df_testing.shape

df_testing['Drop_off'] = 0
df_testing.loc[(df_testing.UNLISTED_REASON == 6) | (df_testing.UNLISTED_REASON == 9) | (df_testing.UNLISTED_REASON == 2) | (df_testing.UNLISTED_REASON == 10) | (df_testing.UNLISTED_REASON == 0), 
               'Drop_off'] = 1  

#6 Medical unsuitable
#9 Other 
#2 Patients expired
#10 Status change
#0 Unknown


df_testing['UNLISTED_DATE'] = pd.to_datetime(df_testing['UNLISTED_DATE'])
df_testing['DATE_OF_LAB'] = pd.to_datetime(df_testing['DATE_OF_LAB'])
df_testing['Interval'] = df_testing['UNLISTED_DATE'] - df_testing['DATE_OF_LAB']
df_testing['Interval'] = df_testing['Interval'].dt.days

df_testing

df_testing['Drop_3M'] = 0
df_testing.loc[(df_testing.Interval <= 90)&(df_testing.Drop_off == 1), 'Drop_3M'] = 1 

df_testing.shape

df_positive = df_testing[df_testing.Interval > 0]
df_positive.shape

df_testing.to_csv('combined_2.csv')

df_testing.to_excel('combined_2.xlsx')
