==> 2.IntegratedModelTesting.ipynb <==
# Integrated File to finish the following task 
# 1. check basic info of the datasets 
# 2. conduct basic data cleansing 
# 3. split the data for training and testing 
# 4. Build the logisitic regression model and conduct cross validation to 
# figure out the best model (use accuracy for now? Need to discusss *0*)
# 5. Build the decision tree model and also use cross validation to figure 
# out the best model 
# 6. Test both models using the testing sets 
# 7. Draw the graph to compare all the exising models(more data to finish)

import numpy as np
import pandas as pd
import seaborn as sn
import matplotlib.pyplot as plt
import seaborn as sns 
import statsmodels.api as sm
sns.set (style = 'white')
sns.set (style = 'whitegrid', color_codes = True)

from sklearn.model_selection import train_test_split

np.set_printoptions(precision=2)

def calc_dynamic(df_val):
    #CREAT
    df_val['PREV_SUBJECT_ID'] = df_val.SUBJECT_ID.shift(1)
    df_val['PREV_CREAT'] = df_val.CREAT.shift(1)
    df_val.PREV_CREAT[ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['PREV2_SUBJECT_ID'] = df_val.PREV_SUBJECT_ID.shift(1)
    df_val['PREV2_CREAT'] = df_val.PREV_CREAT.shift(1)
    df_val.PREV2_CREAT[ df_val['PREV2_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['DEL_CREAT'] = df_val['CREAT'] - (df_val['PREV_CREAT']+df_val['PREV2_CREAT'])/2 
    #BILT
    df_val['PREV_BILT'] = df_val.BILT.shift(1)
    df_val.PREV_BILT[ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['PREV2_BILT'] = df_val.PREV_BILT.shift(1)
    df_val.PREV2_BILT[ df_val['PREV2_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan  
    df_val['DEL_BILT'] = df_val['BILT'] - (df_val['PREV_BILT']+df_val['PREV2_BILT'])/2     
    #AST
    df_val['PREV_AST'] = df_val.AST.shift(1)
    df_val.PREV_AST[ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['PREV2_AST'] = df_val.PREV_AST.shift(1)
    df_val.PREV2_AST[ df_val['PREV2_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['DEL_AST'] = df_val['AST'] - (df_val['PREV_AST']+df_val['PREV2_AST'])/2 
    #ALT
    df_val['PREV_ALT'] = df_val.ALT.shift(1)
    df_val.PREV_ALT[ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['PREV2_ALT'] = df_val.PREV_ALT.shift(1)
    df_val.PREV2_ALT[ df_val['PREV2_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['DEL_ALT'] = df_val['ALT'] - (df_val['PREV_ALT']+df_val['PREV2_ALT'])/2 
    #ALK_PHOS 
    df_val['PREV_ALK_PHOS'] = df_val.ALK_PHOS.shift(1)
    df_val.PREV_ALK_PHOS[ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['PREV2_ALK_PHOS'] = df_val.PREV_ALK_PHOS.shift(1)
    df_val.PREV2_ALK_PHOS[ df_val['PREV2_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['DEL_ALK_PHOS'] = df_val['ALK_PHOS'] - (df_val['PREV_ALK_PHOS']+df_val['PREV2_ALK_PHOS'])/2 
    #PLATELETS
    df_val['PREV_PLATELETS'] = df_val.PLATELETS .shift(1)
    df_val.PREV_PLATELETS [ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['PREV2_PLATELETS'] = df_val.PREV_PLATELETS .shift(1)
    df_val.PREV2_PLATELETS [ df_val['PREV2_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['DEL_PLATELETS'] = df_val['PLATELETS'] - (df_val['PREV_PLATELETS']+df_val['PREV2_PLATELETS'])/2 
    #WBC
    df_val['PREV_WBC'] = df_val.WBC.shift(1)
    df_val.PREV_WBC[ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['PREV2_WBC'] = df_val.PREV_WBC.shift(1)
    df_val.PREV2_WBC[ df_val['PREV2_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['DEL_WBC'] = df_val['WBC'] - (df_val['PREV_WBC']+df_val['PREV2_WBC'])/2 
    #ALB 
    df_val['PREV_ALB'] = df_val.ALB.shift(1)
    df_val.PREV_ALB[ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['PREV2_ALB'] = df_val.PREV_ALB.shift(1)
    df_val.PREV2_ALB[ df_val['PREV2_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['DEL_ALB'] = df_val['ALB'] - (df_val['PREV_ALB']+df_val['PREV2_ALB'])/2 
    #MELD 
    df_val['PREV_MELD'] = df_val.MELD.shift(1)
    df_val.PREV_MELD[ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['PREV2_MELD'] = df_val.PREV_MELD.shift(1)
    df_val.PREV2_MELD[ df_val['PREV2_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['DEL_MELD'] = df_val['MELD'] - (df_val['PREV_MELD']+df_val['PREV2_MELD'])/2 
    #HGB 
    df_val['PREV_HGB'] = df_val.HGB.shift(1)
    df_val.PREV_HGB[ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['PREV2_HGB'] = df_val.PREV_HGB.shift(1)
    df_val.PREV2_HGB[ df_val['PREV2_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['DEL_HGB'] = df_val['HGB'] - (df_val['PREV_HGB']+df_val['PREV2_HGB'])/2 
    
    #INR
    df_val['PREV_INR'] = df_val.INR.shift(1)
    df_val.PREV_INR[ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['PREV2_INR'] = df_val.PREV_INR.shift(1)
    df_val.PREV2_INR[ df_val['PREV2_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['DEL_INR'] = df_val['INR'] - (df_val['PREV_INR']+df_val['PREV2_INR'])/2 
    #Na
    df_val['PREV_Na'] = df_val.Na.shift(1)
    df_val.PREV_Na[ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['PREV2_Na'] = df_val.PREV_Na.shift(1)
    df_val.PREV2_Na[ df_val['PREV2_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['DEL_Na'] = df_val['Na'] - (df_val['PREV_Na']+df_val['PREV2_Na'])/2 
    #HCO3
    df_val['PREV_HCO3'] = df_val.HCO3.shift(1)
    df_val.PREV_HCO3[ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['PREV2_HCO3'] = df_val.PREV_HCO3.shift(1)
    df_val.PREV2_HCO3[ df_val['PREV2_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['DEL_HCO3'] = df_val['HCO3'] - (df_val['PREV_HCO3']+df_val['PREV2_HCO3'])/2 
    #GLUCOSE
    df_val['PREV_GLUCOSE'] = df_val.GLUCOSE.shift(1)
    df_val.PREV_GLUCOSE[ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['PREV2_GLUCOSE'] = df_val.PREV_GLUCOSE.shift(1)
    df_val.PREV2_GLUCOSE[ df_val['PREV2_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['DEL_GLUCOSE'] = df_val['GLUCOSE'] - (df_val['PREV_GLUCOSE']+df_val['PREV2_GLUCOSE'])/2 
    #CALC
    df_val['PREV_CALC'] = df_val.CALC.shift(1)
    df_val.PREV_CALC[ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['PREV2_CALC'] = df_val.PREV_CALC.shift(1)
    df_val.PREV2_CALC[ df_val['PREV2_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['DEL_CALC'] = df_val['CALC'] - (df_val['PREV_CALC']+df_val['PREV2_CALC'])/2 
    #MG
    df_val['PREV_MG'] = df_val.MG.shift(1)
    df_val.PREV_MG[ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['PREV2_MG'] = df_val.PREV_MG.shift(1)
    df_val.PREV2_MG[ df_val['PREV2_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['DEL_MG'] = df_val['MG'] - (df_val['PREV_MG']+df_val['PREV2_MG'])/2 
    #PHOS 
    df_val['PREV_PHOS'] = df_val.PHOS.shift(1)
    df_val.PREV_PHOS[ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['PREV2_PHOS'] = df_val.PREV_PHOS.shift(1)
    df_val.PREV2_PHOS[ df_val['PREV2_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = np.nan
    df_val['DEL_PHOS'] = df_val['PHOS'] - (df_val['PREV_PHOS']+df_val['PREV2_PHOS'])/2 
    
    return df_val

# Function to build logostic regression model
def build_lr(X_train,y_train):

    from sklearn.model_selection import GridSearchCV
    from sklearn.linear_model import LogisticRegression
    grid={"C":np.logspace(-5,5,11), "penalty":["l1","l2"]}# l1 lasso l2 ridge
    logreg=LogisticRegression()
    logreg_cv=GridSearchCV(logreg,grid,cv=5,scoring='roc_auc')
    #need to look at scoring criteria 
    
    clf =logreg_cv.fit(X_train,y_train)
    clf.score(X_train, y_train)
    print("tuned hpyerparameters :(best parameters) ",clf.best_params_)
    print("accuracy :",clf.best_score_)
    
    return clf

# Decision tree model
def build_dct(X_train,y_train):

    import graphviz 
    from sklearn import tree
    from sklearn.model_selection import train_test_split
    from sklearn.grid_search import GridSearchCV
    sample_split_range = list(range(2, 50))
    param_grid = dict(min_samples_split=sample_split_range)
    
    # instantiate the grid
    dtc = tree.DecisionTreeClassifier(max_depth = 5)
    dtc_cv = GridSearchCV(dtc, param_grid, cv=5, scoring='roc_auc')
    clf_dct = dtc_cv.fit(X_train, y_train)
    clf_dct = dtc_cv.best_estimator_
    
    # Draw the decision tree 
    dot_data = tree.export_graphviz(clf_dct, out_file=None, 
                         feature_names=feature_patients1,  
                         class_names=target_states,  
                         filled=True, rounded=True,  
                          special_characters=True)  
    graph = graphviz.Source(dot_data)  
    graph.format = 'png'
    graph.render('Decision Tree',view=True)

    return clf_dct

#Building the random forest model 
############################### still needs review  
def build_dct(X_train,y_train):
    from sklearn.ensemble import RandomForestClassifier
    from sklearn.metrics import mean_squared_error
    from sklearn.model_selection import cross_val_score
    from sklearn import metrics
    from sklearn.model_selection import GridSearchCV

    params_to_search={'n_estimators':[10,50,100,150],
                      'max_depth':[6,8,10,12],
                      'min_samples_split':[2,5,8,10],
                      'min_samples_leaf':[2,5,8,10],
                     }
    mdlRF = RandomForestClassifier(criterion='gini',bootstrap=True)
    optimized_clf = GridSearchCV(mdlRF,params_to_search,cv=5)
    optimized_clf.fit(xTrain,yTrain)

    score = optimized_clf.best_score_
    print(score)
    optimized_clf.best_params_

    #yPred = optimized_clf.predict(xTest)
    #scores = cross_val_score(, x_train, y_train, cv=5)
    #mean_squared_error(yTest, yPred)

    mdlAcc_rf = np.mean((clf.predict(X_train) == y_train))

    return mdlRF
    # -------------------
########################################


df = pd.read_csv('combined_2.csv')

# Drop unnecessary columns 

df_val = df[:]
df_val = df_val.sort_values(by=['SUBJECT_ID'], ascending = True )
# convert the unit of Billirubin and Albumin 
df_val['BILT'] = df_val['BILT']/17.1
df_val['ALB'] = df_val['ALB']/10


df_val.head()

# Creat interval variables 
df_val = calc_dynamic(df_val)

# listing time
df_val['LISTING_TIME'] =pd.to_datetime(df_val['DATE_OF_LAB']) - pd.to_datetime(df_val['LISTING_DATE'])
df_val['LISTING_TIME'] = (df_val['LISTING_TIME'].dt.days)/365
df_val['Age'] = df_val['AGE_AT_LISTING'] +  df_val['LISTING_TIME']

df_val.drop(['BW_1mo','BW_2mo','BW_3mo','BW_4mo','BW_5mo','BW_6mo','BW_7mo','BW_8mo','BW_9mo',
             'BW_10mo','BW_11mo','BW_1yr','BW_>1yr','Unlisted_Lab_Int','Identifier','Identifier.1','Identifier'], inplace=True, axis=1, errors='ignore')
df_val.drop(['SUBJECT_ID.2','LISTING_DATE.2','DATE_OF_LAB.1','Identifier','BW At Listing','SUBJECT_ID.1','LISTING_DATE.1'],inplace=True, axis=1, errors='ignore')
df_val.drop(['DIALYSIS_TYPE_DESCR'], inplace=True, axis=1, errors='ignore')
df_val.drop(['DATE_OF_TRANSPLANT','AGE_AT_UNLISTING','AGE_AT_DEATH','DATE_OF_LAB','LAB_INTERVAL',
             'Identifier','LAB_INTERVAL.1'], inplace=True, axis=1, errors='ignore')
#df_val.drop(['EST_GFR'], inplace=True, axis=1, errors='ignore')
df_val.drop(['RACE', 'RACE_DESCR'], inplace=True, axis=1, errors='ignore')

df_val.head()

# calculate EST_GFR
# check the categorical features 
#print(df.groupby('RACE_DESCR').RACE_DESCR.count())
#print(df_val.groupby('BLOOD_TYPE_DESCR').BLOOD_TYPE_DESCR.count())
#print(df_val.groupby('RH_FACTOR_DESCR').RH_FACTOR_DESCR.count())
#print(df_val.groupby('P_DIAG_DESCR').P_DIAG_DESCR.count())
#print(df_val.groupby('SEC_DIAG_DESCR').SEC_DIAG_DESCR.count())
#print(df_val.groupby('SMOKER').SMOKER.count())
#print(df_val.groupby('PREV_SURG1').PREV_SURG1.count())
# Convert Categorical features to Binary 
#data_race = pd.get_dummies(df_val['RACE_DESCR'])

data_blood = pd.get_dummies(df_val['BLOOD_TYPE_DESCR'])
RH_map = {"Positive": 1, "Negative": 0}
Dialysis_map = {"1": 1,"2": 1,"3": 1,"6": 1,"0": 0}
df_val.replace(RH_map, inplace = True)
df_val.DIALYSIS_TYPE[df_val['DIALYSIS_TYPE']!=0] = 1
df_val['PREV_DIALYSIS_TYPE'] = df_val.DIALYSIS_TYPE.shift(1)
df_val.PREV_DIALYSIS_TYPE[ df_val['PREV_SUBJECT_ID'] != df_val['SUBJECT_ID'] ] = 0
df_val = pd.concat([df_val,data_blood],axis = 1, sort = False)

df_val.rename(columns={"PREV_SURG1": "PREV_SURG"}, inplace=True)

# calculating the log values]
df_val['eBILT'] = np.log10(df_val['BILT'])
df_val['eCREAT'] = np.log10(df_val['CREAT'])
df_val['eINR'] = np.log10(df_val['INR'])

list(df_val)

## Export the data here for OPOM validation 
df_val.to_csv('OPOM_testing.csv')

list(df_val)

# Order the features 
# Select Features for testing 
#feature_patients_all = ['CREAT',	'BILT',	'AST',	'ALT',	'ALK_PHOS',	'PLATELETS',	'WBC',	'ALB',	'MELD',	'HGB',	'INR',	'AGE_AT_LISTING',	'PATIENT_STATUS',	'TREATMENT_PHASE',	'SEX',	'RACE',	'HEIGHT_AT_LISTING',	'WEIGHT_AT_LISTING',	'BLOOD_TYPE',	'RH_FACTOR',	'P_DIAG',	'SEC_DIAG',	'SMOKER',	'PREV_SURG',	'DIALYSIS_TYPE',	'Na',	'HCO3',	'GLUCOSE',	'CALC',	'MG',	'PHOS','LISTING_TIME','Drop_3M']

# We will ignore diagnosis for now 
feature_testing1 = ['DEL_CREAT','CREAT','DEL_BILT','BILT','DEL_AST','AST','DEL_ALT','ALT','DEL_PLATELETS','PLATELETS' ,'DEL_WBC','WBC', 'DEL_ALB','ALB', 'DEL_MELD','MELD', 'DEL_HGB','HGB', 'DEL_INR','INR', 'AGE_AT_LISTING', 'HEIGHT_AT_LISTING',	'WEIGHT_AT_LISTING','RH_FACTOR_DESCR','SMOKER',	'PREV_SURG',	'DIALYSIS_TYPE',	'DEL_Na','Na','DEL_HCO3','HCO3','DEL_GLUCOSE','GLUCOSE','DEL_CALC','CALC','DEL_MG','MG','DEL_PHOS','PHOS','LISTING_TIME','A','AB','O','B','Drop_3M']
#feature_testing2 = ['CREAT','BILT','AST','ALT','PLATELETS', 'WBC', 'ALB', 'MELD', 'HGB', 'INR', 'AGE_AT_LISTING', 'HEIGHT_AT_LISTING',	'WEIGHT_AT_LISTING','RH_FACTOR_DESCR','SMOKER',	'PREV_SURG',	'DIALYSIS_TYPE',	'Na',		'LISTING_TIME','A','AB','O','B','Drop_3M']
#feature_testing1 = ['DEL_CREAT','CREAT','DEL_BILT','BILT','DEL_AST','AST','DEL_ALT','ALT','DEL_PLATELETS','PLATELETS' ,'DEL_WBC','WBC', 'DEL_ALB','ALB', 'DEL_MELD','MELD', 'DEL_HGB','HGB', 'DEL_INR','INR', 'AGE_AT_LISTING', 'HEIGHT_AT_LISTING',	'WEIGHT_AT_LISTING','RH_FACTOR_DESCR','SMOKER',	'PREV_SURG',	'DIALYSIS_TYPE',	'DEL_Na','Na','LISTING_TIME','Asian','Black','Caucasian','First Nations', 'Indian Sub-Continent', 'Indigenous/Aboriginal/Inuit not included elsewhere','Latin American','Mid-East/Arabian','Other','Unknown','A','AB','O','B','Drop_3M']


df_val = df_val[feature_testing1]


# quick look at the data 
#table1 = np.mean(patients, axis = 0)
#table2 = np.std(patients, axis = 0 )

#feature_patients = ['CREAT','BILT','AST','ALT','PLATELETS', 'WBC', 'ALB', 'MELD', 'HGB', 'INR', 'AGE_AT_LISTING', 'HEIGHT_AT_LISTING',	'WEIGHT_AT_LISTING','RH_FACTOR_DESCR','SMOKER',	'PREV_SURG',	'DIALYSIS_TYPE',	'Na',	'HCO3',	'GLUCOSE',	'CALC',	'MG',	'PHOS','LISTING_TIME','Asian','Black','Caucasian','First Nations', 'Indian Sub-Continent', 'Indigenous/Aboriginal/Inuit not included elsewhere','Latin American','Mid-East/Arabian','Other','Unknown','A','AB','O','B']
#feature_patients2 = ['CREAT','BILT','AST','ALT','PLATELETS', 'WBC', 'ALB', 'MELD', 'HGB', 'INR', 'AGE_AT_LISTING', 'HEIGHT_AT_LISTING',	'WEIGHT_AT_LISTING','RH_FACTOR_DESCR','SMOKER',	'PREV_SURG',	'DIALYSIS_TYPE',	'Na',	'LISTING_TIME','A','AB','O','B']
feature_patients1 = ['DEL_CREAT','CREAT','DEL_BILT','BILT','DEL_AST','AST','DEL_ALT','ALT','DEL_PLATELETS','PLATELETS' ,'DEL_WBC','WBC', 'DEL_ALB','ALB', 'DEL_MELD','MELD', 'DEL_HGB','HGB', 'DEL_INR','INR', 'AGE_AT_LISTING', 'HEIGHT_AT_LISTING',	'WEIGHT_AT_LISTING','RH_FACTOR_DESCR','SMOKER',	'PREV_SURG',	'DIALYSIS_TYPE',	'DEL_Na','Na','DEL_HCO3','HCO3','DEL_GLUCOSE','GLUCOSE','DEL_CALC','CALC','DEL_MG','MG','DEL_PHOS','PHOS','LISTING_TIME','A','AB','O','B']
#feature_patients = ['CREAT',	'BILT',	'AST',	'ALT',	'ALK_PHOS',	'PLATELETS',	'WBC',	'ALB',	'MELD',	'HGB',	'INR',	'AGE_AT_LISTING',	'PATIENT_STATUS',	'TREATMENT_PHASE',	'SEX',	'RACE',	'HEIGHT_AT_LISTING',	'WEIGHT_AT_LISTING',	'BLOOD_TYPE',	'RH_FACTOR',	'P_DIAG',	'SEC_DIAG',	'SMOKER',	'PREV_SURG',	'DIALYSIS_TYPE',	'Na',	'HCO3',	'GLUCOSE',	'CALC',	'MG',	'PHOS','LISTING_TIME']
#feature_patients_backup = ['CREAT',	'BILT',	'AST',	'ALT', 'PLATELETS',	'WBC',	'ALB','MELD','HGB',	'INR',	'AGE_AT_LISTING',	'PATIENT_STATUS',	'TREATMENT_PHASE','P_DIAG',	'SEC_DIAG','PREV_SURG',	'DIALYSIS_TYPE']
#feature_patients1 = ['DEL_CREAT','CREAT','DEL_BILT','BILT','DEL_AST','AST','DEL_ALT','ALT','DEL_PLATELETS','PLATELETS' ,'DEL_WBC','WBC', 'DEL_ALB','ALB', 'DEL_MELD','MELD', 'DEL_HGB','HGB', 'DEL_INR','INR', 'AGE_AT_LISTING', 'HEIGHT_AT_LISTING',	'WEIGHT_AT_LISTING','RH_FACTOR_DESCR','SMOKER',	'PREV_SURG',	'DIALYSIS_TYPE',	'DEL_Na','Na','LISTING_TIME','Asian','Black','Caucasian','First Nations', 'Indian Sub-Continent', 'Indigenous/Aboriginal/Inuit not included elsewhere','Latin American','Mid-East/Arabian','Other','Unknown','A','AB','O','B']

## Data Exploration
patients = df_val.dropna(how='any')
patients.info()
patients['Drop_3M'].value_counts()

#sns.countplot (x = 'Drop_3M', data= patients, palette = 'hls')
#plt.show()
#plt.savefig('count_plot.png')

X_patients = patients[feature_patients1]
y_patients = patients['Drop_3M']
#target_states = ['Drop_off', 'Live']

###Visualization 

#table = pd.crosstab(patients['Na'] ,patients['Drop_3M'])
#table.div(table.sum(1).astype(float), axis = 0).plot(kind = 'bar', stacked = True)
#plt.title('State of Patient for Na')
#plt.xlabel('Na')
#plt.ylabel('State')
#plt.savefig('Na_states.png')

#see correlation between features 
# corr_table = X_patients.corr()  


#Spliting the training set and test set 
X_train, X_test, y_train, y_test = train_test_split(X_patients, y_patients, test_size = 0.3, random_state=5)



# Building Logistic Regression Models 

clf = build_lr(X_train,y_train)

#Stats summary for the model 
mydata = {}
mydata['x'] = X_train
mydata['y'] = y_train 

import statsmodels.formula.api as smf
 

mylogit = smf.glm(formula='y ~ x', data=mydata, family=sm.families.Binomial())
res = mylogit.fit()
print(res.summary())



#Predicting the test set results and calculating the acurracy
y_lr_pred = clf.predict_proba(X_test)
y_lr_pred_rounded = clf.predict(X_test)
print ('Accuracy of logisitic regression classifier on the test set:{:.5f}'.format
           (clf.score(X_test, y_test)))

# Calculate ROC, AUC
from sklearn.metrics import roc_auc_score
from sklearn.metrics import roc_curve
fpr_lr, tpr_lr, threshold_lr = roc_curve( y_test,y_lr_pred[:,1], pos_label = 1)
AUC_lr = roc_auc_score( y_test,y_lr_pred[:,1])
print('AUC for logistic regression model is :', AUC_lr)


# Building the Decision Tree Model 
clf_dct = build_dct(X_train,y_train)

# Calculate y_pred
y_pred_dct = clf_dct.predict_proba(X_test)
y_pred_dct_rounded = clf_dct.predict(X_test)

print('Accuracy of Decision Tree classifier on test set: {:.2f}'
     .format(clf_dct.score(X_test, y_test)))

from sklearn.metrics import roc_auc_score
from sklearn.metrics import roc_curve

#logit1.fit(inputData,outputData)
clf_dct.score(X_test, y_test)
##Computing false and true positive rates
fpr_dct, tpr_dct, thresholds_dct = roc_curve( y_test,y_pred_dct[:,1], drop_intermediate=False)
#clf_dct.predict(X_test)

AUC_dct = roc_auc_score( y_test,y_pred_dct[:,1])

print('AUC for Decision Tree Model is :', AUC_dct)
