==> 3.OPOM-Data-Preparation.ipynb <==
# Import packages
import matplotlib
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd

df_OPOM = pd.read_csv('OPOM_testing.csv')

list(df_OPOM)


df_OPOM = df_OPOM.rename(columns={"PREV_CREAT":"pCREAT", 'PREV_BILT': 'pBILT', 'PREV_ALB': 'pALB', 'PREV_MELD': 'pMELD', 'PREV_INR': 'pINR', 'PREV_MELD': 'pMELD', 
                                  'PREV_Na': 'pNa', 'DEL_BILT':'dBILT',  'DEL_ALB':'dALB', 'DEL_INR':'dINR',  'DEL_MELD':'dMELD', 'DEL_Na':'dNa',
                                  'LISTING_TIME': 'ListYears','DIALYSIS_TYPE': 'Dialysis', 'PREV_DIALYSIS_TYPE': 'pDialysis'})

df_OPOM['Live'] = 1- df_OPOM['Drop_3M']


OPOM_columns = ['CREAT', 'BILT', 'ALB', 'MELD', 'INR', 'Na' ,'pCREAT', 'pBILT', 'pALB', 'pMELD', 'pINR', 'pMELD', 'pNa', 'pCREAT', 'dBILT',  'dALB', 'dINR',  'dMELD', 'dNa',   'Age',  
                'ListYears', 'Dialysis', 'pDialysis','eBILT', 'eCREAT', 'eINR', 'SUBJECT_ID', 'Live', 'P_DIAG_DESCR', 'SEC_DIAG_DESCR', 'Drop_3M'
               ]

df_OPOM1 = df_OPOM[OPOM_columns]

#df_val.DIALYSIS_TYPE[df_val['DIALYSIS_TYPE']!=0] = 1
df_OPOM_NHCC = df_OPOM1[df_OPOM1['P_DIAG_DESCR'] != 'Hepatocellular Ca (HCC) ']

df_OPOM_NHCC2 = df_OPOM_NHCC[df_OPOM_NHCC['SEC_DIAG_DESCR'] != 'Hepatocellular Ca (HCC) ']

OPOM_patients = df_OPOM_NHCC2.dropna(how='any')

OPOM_patients.MELD.head()

# Calculate mortality based on MELD
OPOM_patients.loc[(OPOM_patients.MELD >= 40), 'MELD_prob'] = 0.713
OPOM_patients.loc[(OPOM_patients.MELD <= 39)&(OPOM_patients.MELD >= 30), 'MELD_prob'] = 0.526
OPOM_patients.loc[(OPOM_patients.MELD <= 29)&(OPOM_patients.MELD >= 20), 'MELD_prob'] = 0.196
OPOM_patients.loc[(OPOM_patients.MELD <= 19)&(OPOM_patients.MELD >= 10), 'MELD_prob'] = 0.06
OPOM_patients.loc[(OPOM_patients.MELD <= 9), 'MELD_prob'] = 0.019

# Calculate MELD-NA, rounded to nearest int
OPOM_patients['MELD_Na'] = round(OPOM_patients['MELD'] - OPOM_patients['Na'] - (0.025*OPOM_patients['MELD']*(140-OPOM_patients['Na']))+140, 0)

# Calculate mortality based on MELD_Na
OPOM_patients.loc[OPOM_patients.MELD_Na >= 40, 'MELD_Na_prob'] = 0.713
OPOM_patients.loc[(OPOM_patients.MELD_Na <= 39)&(OPOM_patients.MELD_Na >= 30), 'MELD_Na_prob'] = 0.526
OPOM_patients.loc[(OPOM_patients.MELD_Na <= 29)&(OPOM_patients.MELD_Na >= 20), 'MELD_Na_prob'] = 0.196
OPOM_patients.loc[(OPOM_patients.MELD_Na <= 19)&(OPOM_patients.MELD_Na >= 10), 'MELD_Na_prob'] = 0.06
OPOM_patients.loc[OPOM_patients.MELD_Na <= 9, 'MELD_Na_prob'] = 0.019

a = OPOM_patients.SUBJECT_ID.unique()

#len(a)

OPOM_patients.head()

OPOM_patients.to_csv('OPOM_tested.csv')
